Watchdog: fix a number of corner cases.

+++ src/rp2040/pio.h
@@ -12,6 +12,6 @@ void init_pio(const uint32_t joint);
 uint8_t do_steps(const uint8_t joint, const uint32_t update_time_us);
 
 /* One bit per axis to indicate whether a corresponding PIO has been updated */
-extern uint32_t axis_updated_bitmask;
+extern uint8_t axis_updated_bitmask;
 
 #endif  // PIO__H
